some cleanup

+++ src/py21cmfast/src/HaloBox.c
@@ -878,12 +878,12 @@ int test_halo_props(double redshift, float *vcb_grid, float *J21_LW_grid, float 
                 if (m == 0.) {
                     continue;
                 }
-                x = halo_coords[0 + 3 * i_halo] * simulation_options_global->HII_DIM /
-                    simulation_options_global->BOX_LEN;
-                y = halo_coords[1 + 3 * i_halo] * simulation_options_global->HII_DIM /
-                    simulation_options_global->BOX_LEN;
-                z = halo_coords[2 + 3 * i_halo] * simulation_options_global->HII_DIM /
-                    simulation_options_global->BOX_LEN;
+                x = (int)(halo_coords[0 + 3 * i_halo] * simulation_options_global->HII_DIM /
+                          simulation_options_global->BOX_LEN);
+                y = (int)(halo_coords[1 + 3 * i_halo] * simulation_options_global->HII_DIM /
+                          simulation_options_global->BOX_LEN);
+                z = (int)(halo_coords[2 + 3 * i_halo] * simulation_options_global->HII_DIM /
+                          simulation_options_global->BOX_LEN);
                 i_cell = HII_R_INDEX(x, y, z);
 
                 // set values before reionisation feedback
